Implemented Title Scene
Implemented a basic title scene with a SceneHandler that will allow the player to navigate to the main scene and back to the title scene. While "test" is selected in the SceneHandler object's inspector, the title scene will automatically be loaded after a couple of seconds of being in the main scene.